Move Partition_Tracker into own file
