+++ src/pegr/algs/Partition_Tracker.hpp
@@ -1,0 +1,39 @@
+/*
+ *  Copyright 2017 James Fong
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+#ifndef PEGR_ALGS_PARTITIONTRACKER_HPP
+#define PEGR_ALGS_PARTITIONTRACKER_HPP
+
+#include <vector>
+
+namespace pegr {
+namespace Algs {
+
+class Partition_Tracker {
+public:
+    void occupy(std::size_t offset, std::size_t size);
+    
+    bool can_occupy(std::size_t offset, std::size_t size);
+    
+    std::size_t get_minimum_size();
+private:
+    std::vector<char> m_data;
+};
+
+} // namespace Algs
+} // namespace pegr
+
+#endif // PEGR_ALGS_PARTITIONTRACKER_HPP
